누락된 기능 추가 및 전체적인 리팩토링

+++ src/main/java/com/api/farmingsoon/domain/bid/service/BidService.java
@@ -1,5 +1,6 @@
 package com.api.farmingsoon.domain.bid.service;
 
+import com.api.farmingsoon.common.event.BidRegisterEvent;
 import com.api.farmingsoon.common.exception.ErrorCode;
 import com.api.farmingsoon.common.exception.custom_exception.NotFoundException;
 import com.api.farmingsoon.common.util.AuthenticationUtils;
@@ -7,48 +8,55 @@ import com.api.farmingsoon.domain.bid.dto.BidRequest;
 import com.api.farmingsoon.domain.bid.dto.BidWithPageResponse;
 import com.api.farmingsoon.domain.bid.dto.BidsResponse;
 import com.api.farmingsoon.domain.bid.model.Bid;
+import com.api.farmingsoon.domain.bid.model.BidResult;
 import com.api.farmingsoon.domain.bid.repository.BidRepository;
 import com.api.farmingsoon.domain.item.domain.Item;
 import com.api.farmingsoon.domain.item.repository.ItemRepository;
 import com.api.farmingsoon.domain.member.model.Member;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+
 @Service
-@Transactional
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class BidService {
 
     private final BidRepository bidRepository;
     private final ItemRepository itemRepository;
     private final AuthenticationUtils authenticationUtils;
+    private final ApplicationEventPublisher eventPublisher;
 
+    @Transactional
     public void bid(BidRequest bidRequest) {
         Member member = authenticationUtils.getAuthenticationMember();
         Item item = itemRepository.findById(bidRequest.getItemId()).orElseThrow(() -> new NotFoundException(ErrorCode.NOT_FOUND_ITEM));
 
-        bidRepository.save(Bid.of(item, member, bidRequest.getPrice()));
+        bidRepository.save(Bid.of(item, member, bidRequest.getPrice(), BidResult.BIDDING));
+        eventPublisher.publishEvent(new BidRegisterEvent(item)); // 아이템 판매자에게 알림전송
     }
 
-    @Transactional(readOnly = true)
-    public BidWithPageResponse myBidList(Pageable pageable) {
-        Member member = authenticationUtils.getAuthenticationMember();
-
-        return BidWithPageResponse.of(bidRepository.findAllByMember(member, pageable));
-    }
-
+    @Transactional
     public void delete(Long bidId) {
-        Member member = authenticationUtils.getAuthenticationMember();
         Bid bid = bidRepository.findById(bidId).orElseThrow(() -> new NotFoundException(ErrorCode.NOT_FOUND_BID));
+        AuthenticationUtils.checkDeletePermission(bid.getMember());
 
         bidRepository.delete(bid);
     }
 
+    @Transactional
     public void deleteAll() {
         Member member = authenticationUtils.getAuthenticationMember();
         bidRepository.deleteAllByMember(member);
+    }
+
+
+    public Page<Bid> getMyBidList(Member member, Pageable pageable){
+        return bidRepository.findAllByMember(member, pageable);
     }
 }
